requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.0.1
python-telegram-bot==20.7
qrcode==7.4.2
requests==2.31.0